+++ src/components/ConstructorPizza.js
@@ -1,5 +1,4 @@
 import React from "react";
-import ButtonComponet from "./UI/button/ButtonComponet";
 
 const ConstructorPizza = (props) => {
     return (
