+++ src/components/UI/button/ButtonComponent.js
@@ -2,7 +2,7 @@ import React from "react";
 import Button from "@atlaskit/button";
 import "./btn.scss";
 
-const ButtonComponet = ({onClick, buttonName, buttonType}) => {
+export const ButtonComponent = ({onClick, buttonName, buttonType}) => {
     return (
         <div className='btn'>
             <Button onClick={onClick} className='btn' type={buttonType}>{buttonName}</Button>
@@ -11,4 +11,4 @@ const ButtonComponet = ({onClick, buttonName, buttonType}) => {
 }
 
 
-export default ButtonComponet;
+// export default ButtonComponet;
