+++ src/components/UI/cards/DiscountCard.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-const DiscountCard = (props) => {
+export const DiscountCard = (props) => {
     const {img, listTitle} = props;
     return (
         <div className="discount-item">
@@ -13,4 +13,4 @@ const DiscountCard = (props) => {
     );
 }
 
-export default DiscountCard;
+// export default DiscountCard;
