+++ src/components/UI/cards/DrinkCard.js
@@ -1,7 +1,7 @@
 import React from "react";
-import ButtonComponet from "../button/ButtonComponet";
+import {ButtonComponent} from "../index";
 
-const DrinkCard = (props) => {
+export const DrinkCard = (props) => {
     const {img, title, size, price} = props;
     return (
         <div className="product-item drink-item">
@@ -20,10 +20,10 @@ const DrinkCard = (props) => {
                 </div>
             </a>
             <div className="product-item__buy">
-                <ButtonComponet buttonName='Додати в корзину'/>
+                <ButtonComponent buttonName='Додати в корзину'/>
             </div>
         </div>
     );
 }
 
-export default DrinkCard;
+// export default DrinkCard;
