+++ src/components/UI/index.js
@@ -1,13 +1,13 @@
-export ButtonComponet from "./button/ButtonComponet";
-export Card from "./cards/Card";
-export DiscountCard from "./cards/DiscountCard";
-export DrinkCard from "./cards/DrinkCard";
+export {ButtonComponent} from "./button/ButtonComponent";
+export {Card} from "./cards/Card";
+export {DiscountCard} from "./cards/DiscountCard";
+export {DrinkCard} from "./cards/DrinkCard";
 
-export SliderAlso from "./carousel/SliderAlso";
+export {SliderAlso} from "./carousel/SliderAlso";
 
-export InputComponet from "./input/inputComponent";
+export {InputComponent} from "./input/inputComponent";
 
-export ModalComponent from "./modal/ModalComponent";
-export ModalBasket from "./modal/ModalBasket";
-export BasketForm from "./modal/BasketForm";
-export SmallCard from "./modal/SmallCard";
+export {ModalComponent} from "./modal/ModalComponent";
+export {ModalBasket} from "./modal/ModalBasket";
+export {BasketForm} from "./modal/BasketForm";
+export {SmallCard} from "./modal/SmallCard";
